requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Evaluation
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
sentence-transformers>=2.2.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Quantization
llama-cpp-python>=0.2.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
psutil>=5.9.0
GPUtil>=1.4.0

# Development
pytest>=7.4.0
black>=23.11.0
ruff>=0.1.5
ipykernel>=6.26.0
jupyter>=1.0.0
matplotlib>=3.8.0
